refactor: use correct naming for mods
